Uncaught (in promise) TypeError: Cannot set properties of null (setting 'innerHTML')
