+++ index.html
@@ -8,7 +8,10 @@
 </head>
 
 <body>
-  <div id="renderAll"></div>
+  <div class="container" id="container">
+    <ul class="comments" id="commentsId"></ul>
+    <div id="user-input"></div>
+  </div>
   <!-- <div class="add-form">
     <input type="text" class="add-form-name" id="nameFormId" placeholder="Введите ваше имя" />
     <textarea type="textarea" class="add-form-text" id="textFormId" placeholder="Введите ваш коментарий"
