+++ renderComments.js
@@ -1,8 +1,5 @@
-import { commentPostListener, copyToRespond, initLikeButtonListeners } from "./listeners.js";
+import { copyToRespond, initLikeButtonListeners } from "./listeners.js";
 import { safeInput } from "./helpers.js";
-import { renderLogin } from "./renderLogin.js";
-
-const renderAllElements = document.getElementById('renderAll');
 
 export let userName;
 
@@ -12,11 +9,11 @@ export const setName = (newUserName) => {
 
 export const renderComments = ({ comments, fetchGet }) => {
 
+    const commentsList = document.getElementById('commentsId');
+
     const commentsHtml = comments.map((comment, index) => {
 
         return `
-        <div class="container">
-        <ul class="comments" id="commentsId">
           <li class="comment" data-index="${index}">
             <div class="comment-header">
               <div id="userName">${safeInput({ text: comment.name })}
@@ -34,51 +31,14 @@ export const renderComments = ({ comments, fetchGet }) => {
               </div>
             </div>
           </li>
-        </ul>
-        <div id="user-input">
-        <button class="add-form-button" id="authorisation-button">Чтобы добавить комментарий, авторизуйтесь</button>
-        </div>
-        </div>
     `}).join('');
 
-    renderAllElements.innerHTML = commentsHtml;
-
-    const authorisationButtonElement = document.getElementById("authorisation-button");
-
-    authorisationButtonElement.addEventListener("click", () => {
-        renderLogin({ comments, fetchGet });
-    })
+    commentsList.innerHTML = commentsHtml;
 
     initLikeButtonListeners({ comments, fetchGet });
     copyToRespond({ comments });
 
-
-
     // const commentsList = document.getElementById('commentsId');
     // commentsList.textContent = 'Пожалуйста подождите, комментарии загружаются...';
 
-
 };
-
-export const renderInputForm = ({ comments, fetchGet }) => {
-
-    renderComments({ comments });
-
-    const userInputHtml = document.getElementById("user-input");
-    // const userNameElement = document.getElementById("userName");
-
-    const inputForm = `
-    <div class="add-form">
-    <input type="text" class="add-form-name" id="nameFormId" placeholder="Введите ваше имя" />
-    <textarea type="textarea" class="add-form-text" id="textFormId" placeholder="Введите ваш коментарий"
-      rows="4"></textarea>
-    <div class="add-form-row">
-      <button class="add-form-button" id="button">Написать</button>
-    </div>
-  </div>`;
-
-    userInputHtml.innerHTML = inputForm;
-
-    commentPostListener({ fetchGet });
-};
-
